refactor: collections code (#29)
