--- beginner/array/example1.go
@@ -1,14 +1,0 @@
-package array
-
-import (
-	"fmt"
-)
-
-var users = [...]string{"Luis", "Jorge", "Oscar"}
-
-func main() {
-	fmt.Println(users)
-	for _, value := range users {
-		fmt.Println(value)
-	}
-}
