--- beginner/array/vectors.go
@@ -1,17 +1,0 @@
-package arrays
-
-import (
-	"fmt"
-
-	"./slices"
-)
-
-var users = [...]string{"Luis", "Chris", "Oscar"}
-
-func main() {
-	fmt.Println(users)
-	for _, value := range users {
-		fmt.Println(value)
-	}
-	slices.P
-}
